EXL-294 - Translated view class names - Moved getMenuContent and getMenu name to View class - Added ro_RO translation files
